pyproject.toml
[project]
name = "video-downloader"
version = "1.0.0"
description = "A modern video downloader with GUI using you-get and ffmpeg"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "openai-whisper>=20231117",
    "customtkinter>=5.2.2",
    "packaging>=26.0",
    "yt-dlp>=2025.10.14",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchvision>=0.15.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core", "ui", "utils"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }
torchvision = { index = "pytorch" }
